requirements.txt
# requirements.txt
typer>=0.9.0
rich>=13.7.0
pyyaml>=6.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
tiktoken